🚧 WIP Added some more UI stuff

+++ WeatherAppGG/WeatherAppGG/Models/Condition.swift
@@ -28,7 +28,7 @@ struct Condition: Codable {
     }
 
     struct Wind: Codable {
-        let speedKts: Int?
+        let speedKts: Double?
         let from: Int?
         let direction: Int?
         let variable: Bool?
